Implement 400,404,401 status handle, Complete 3th hw
