+++ src/services/index.js
@@ -26,6 +26,4 @@ instance.interceptors.response.use(
     }
   )
 
-
-
 export default instance
